% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_operations.R
\name{batch_describe_simulation_job}
\alias{batch_describe_simulation_job}
\title{Describes one or more simulation jobs}
\usage{
batch_describe_simulation_job(
  jobs = list(
    "string"
  )
)
}
\arguments{
\item{jobs}{[required] A list of Amazon Resource Names (ARNs) of simulation jobs to describe.}
}
\description{
Describes one or more simulation jobs.
}
\examples{

}
