% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_operations.R
\name{create_fleet}
\alias{create_fleet}
\title{Creates a fleet, a logical group of robots running the same robot application}
\usage{
create_fleet(name, tags = NULL)
}
\arguments{
\item{name}{[required] The name of the fleet.}

\item{tags}{A map that contains tag keys and tag values that are attached to the fleet.}
}
\description{
Creates a fleet, a logical group of robots running the same robot application.
}
\section{Accepted Parameters}{
\preformatted{create_fleet(
  name = "string",
  tags = list(
    "string"
  )
)
}
}

