% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_operations.R
\name{create_robot}
\alias{create_robot}
\title{Creates a robot}
\usage{
create_robot(name, architecture, greengrassGroupId, tags = NULL)
}
\arguments{
\item{name}{[required] The name for the robot.}

\item{architecture}{[required] The target architecture of the robot.}

\item{greengrassGroupId}{[required] The Greengrass group id.}

\item{tags}{A map that contains tag keys and tag values that are attached to the robot.}
}
\description{
Creates a robot.
}
\section{Accepted Parameters}{
\preformatted{create_robot(
  name = "string",
  architecture = "X86_64"|"ARM64"|"ARMHF",
  greengrassGroupId = "string",
  tags = list(
    "string"
  )
)
}
}

