% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_operations.R
\name{create_robot_application}
\alias{create_robot_application}
\title{Creates a robot application}
\usage{
create_robot_application(
  name = "string",
  sources = list(
    list(
      s3Bucket = "string",
      s3Key = "string",
      architecture = "X86_64"|"ARM64"|"ARMHF"
    )
  ),
  robotSoftwareSuite = list(
    name = "ROS",
    version = "Kinetic"
  )
)
}
\arguments{
\item{name}{[required] The name of the robot application.}

\item{sources}{[required] The sources of the robot application.}

\item{robotSoftwareSuite}{[required] The robot software suite used by the robot application.}
}
\description{
Creates a robot application.
}
\examples{

}
