% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_operations.R
\name{delete_fleet}
\alias{delete_fleet}
\title{Deletes a fleet}
\usage{
delete_fleet(
  fleet = "string"
)
}
\arguments{
\item{fleet}{[required] The Amazon Resource Name (ARN) of the fleet.}
}
\description{
Deletes a fleet.
}
\examples{

}
