% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_operations.R
\name{delete_robot}
\alias{delete_robot}
\title{Deletes a robot}
\usage{
delete_robot(robot)
}
\arguments{
\item{robot}{[required] The Amazon Resource Name (ARN) of the robot.}
}
\description{
Deletes a robot.
}
\section{Accepted Parameters}{
\preformatted{delete_robot(
  robot = "string"
)
}
}

