% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_operations.R
\name{delete_robot_application}
\alias{delete_robot_application}
\title{Deletes a robot application}
\usage{
delete_robot_application(application, applicationVersion = NULL)
}
\arguments{
\item{application}{[required] The Amazon Resource Name (ARN) of the the robot application.}

\item{applicationVersion}{The version of the robot application to delete.}
}
\description{
Deletes a robot application.
}
\section{Accepted Parameters}{
\preformatted{delete_robot_application(
  application = "string",
  applicationVersion = "string"
)
}
}

