% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_operations.R
\name{deregister_robot}
\alias{deregister_robot}
\title{Deregisters a robot}
\usage{
deregister_robot(
  fleet = "string",
  robot = "string"
)
}
\arguments{
\item{fleet}{[required] The Amazon Resource Name (ARN) of the fleet.}

\item{robot}{[required] The Amazon Resource Name (ARN) of the robot.}
}
\description{
Deregisters a robot.
}
\examples{

}
