% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_operations.R
\name{describe_deployment_job}
\alias{describe_deployment_job}
\title{Describes a deployment job}
\usage{
describe_deployment_job(
  job = "string"
)
}
\arguments{
\item{job}{[required] The Amazon Resource Name (ARN) of the deployment job.}
}
\description{
Describes a deployment job. [Does it work regardless of deployment status, e.g. Failed?]
}
\examples{

}
