% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_operations.R
\name{describe_fleet}
\alias{describe_fleet}
\title{Describes a fleet}
\usage{
describe_fleet(
  fleet = "string"
)
}
\arguments{
\item{fleet}{[required] The Amazon Resource Name (ARN) of the fleet.}
}
\description{
Describes a fleet.
}
\examples{

}
