% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_operations.R
\name{describe_robot}
\alias{describe_robot}
\title{Describes a robot}
\usage{
describe_robot(
  robot = "string"
)
}
\arguments{
\item{robot}{[required] The Amazon Resource Name (ARN) of the robot to be described.}
}
\description{
Describes a robot.
}
\examples{

}
