% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_operations.R
\name{describe_robot_application}
\alias{describe_robot_application}
\title{Describes a robot application}
\usage{
describe_robot_application(application, applicationVersion = NULL)
}
\arguments{
\item{application}{[required] The Amazon Resource Name (ARN) of the robot application.}

\item{applicationVersion}{The version of the robot application to describe.}
}
\description{
Describes a robot application.
}
\section{Accepted Parameters}{
\preformatted{describe_robot_application(
  application = "string",
  applicationVersion = "string"
)
}
}

