% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_operations.R
\name{describe_simulation_application}
\alias{describe_simulation_application}
\title{Describes a simulation application}
\usage{
describe_simulation_application(
  application = "string",
  applicationVersion = "string"
)
}
\arguments{
\item{application}{[required] The application information for the simulation application.}

\item{applicationVersion}{The version of the simulation application to describe.}
}
\description{
Describes a simulation application.
}
\examples{

}
