% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_operations.R
\name{list_tags_for_resource}
\alias{list_tags_for_resource}
\title{Lists all tags on a AWS RoboMaker resource}
\usage{
list_tags_for_resource(resourceArn)
}
\arguments{
\item{resourceArn}{[required] The AWS RoboMaker Amazon Resource Name (ARN) with tags to be listed.}
}
\description{
Lists all tags on a AWS RoboMaker resource.
}
\section{Accepted Parameters}{
\preformatted{list_tags_for_resource(
  resourceArn = "string"
)
}
}

