% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_package.R
\docType{package}
\name{paws.robomaker-package}
\alias{paws.robomaker}
\alias{paws.robomaker-package}
\title{paws.robomaker: AWS RoboMaker}
\description{
his section provides documentation for the AWS
    RoboMaker API operations.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
