% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_operations.R
\name{register_robot}
\alias{register_robot}
\title{Registers a robot with a fleet}
\usage{
register_robot(
  fleet = "string",
  robot = "string"
)
}
\arguments{
\item{fleet}{[required] The Amazon Resource Name (ARN) of the fleet.}

\item{robot}{[required] The Amazon Resource Name (ARN) of the robot.}
}
\description{
Registers a robot with a fleet.
}
\examples{

}
