% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_operations.R
\name{restart_simulation_job}
\alias{restart_simulation_job}
\title{Restarts a running simulation job}
\usage{
restart_simulation_job(job)
}
\arguments{
\item{job}{[required] The Amazon Resource Name (ARN) of the simulation job.}
}
\description{
Restarts a running simulation job.
}
\section{Accepted Parameters}{
\preformatted{restart_simulation_job(
  job = "string"
)
}
}

