% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_operations.R
\name{sync_deployment_job}
\alias{sync_deployment_job}
\title{Syncrhonizes robots in a fleet to the latest deployment}
\usage{
sync_deployment_job(
  clientRequestToken = "string",
  fleet = "string"
)
}
\arguments{
\item{clientRequestToken}{[required] Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}

\item{fleet}{[required] The target fleet for the synchronization.}
}
\description{
Syncrhonizes robots in a fleet to the latest deployment. This is helpful if robots were added after a deployment.
}
\examples{

}
