% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_operations.R
\name{update_simulation_application}
\alias{update_simulation_application}
\title{Updates a simulation application}
\usage{
update_simulation_application(application, sources,
  simulationSoftwareSuite, robotSoftwareSuite, renderingEngine,
  currentRevisionId = NULL)
}
\arguments{
\item{application}{[required] The application information for the simulation application.}

\item{sources}{[required] The sources of the simulation application.}

\item{simulationSoftwareSuite}{[required] The simulation software suite used by the simulation application.}

\item{robotSoftwareSuite}{[required] Information about the robot software suite.}

\item{renderingEngine}{[required] The rendering engine for the simulation application.}

\item{currentRevisionId}{The revision id for the robot application.}
}
\description{
Updates a simulation application.
}
\section{Accepted Parameters}{
\preformatted{update_simulation_application(
  application = "string",
  sources = list(
    list(
      s3Bucket = "string",
      s3Key = "string",
      architecture = "X86_64"|"ARM64"|"ARMHF"
    )
  ),
  simulationSoftwareSuite = list(
    name = "Gazebo",
    version = "string"
  ),
  robotSoftwareSuite = list(
    name = "ROS",
    version = "Kinetic"
  ),
  renderingEngine = list(
    name = "OGRE",
    version = "string"
  ),
  currentRevisionId = "string"
)
}
}

