% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_operations.R
\name{create_deployment_job}
\alias{create_deployment_job}
\title{Creates a deployment job}
\usage{
create_deployment_job(
  deploymentConfig = list(
    concurrentDeploymentPercentage = 123,
    failureThresholdPercentage = 123
  ),
  clientRequestToken = "string",
  fleet = "string",
  deploymentApplicationConfigs = list(
    list(
      application = "string",
      applicationVersion = "string",
      launchConfig = list(
        packageName = "string",
        preLaunchFile = "string",
        launchFile = "string",
        postLaunchFile = "string",
        environmentVariables = list(
          "string"
        )
      )
    )
  )
)
}
\arguments{
\item{deploymentConfig}{The requested deployment configuration.}

\item{clientRequestToken}{[required] Unique, case-sensitive identifier that you provide to ensure the idempotency of the request.}

\item{fleet}{[required] The Amazon Resource Name (ARN) of the fleet to deploy.}

\item{deploymentApplicationConfigs}{[required] The deployment application configuration.}
}
\description{
Creates a deployment job.
}
\examples{

}
