% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.robomaker_operations.R
\name{create_fleet}
\alias{create_fleet}
\title{Creates a fleet, a logical group of robots running the same robot application}
\usage{
create_fleet(
  name = "string"
)
}
\arguments{
\item{name}{[required] The name of the fleet.}
}
\description{
Creates a fleet, a logical group of robots running the same robot application.
}
\examples{

}
